+++ src/core/config/config.interfaces.ts
@@ -10,6 +10,7 @@ export interface StartupConfig {
   }
   cspWsOverride?: string
   debug?: boolean
+  webroot?: string
 }
 
 interface PluginChildBridge {
@@ -70,6 +71,7 @@ export interface HomebridgeUiConfig {
   name: string
   port: number
   host?: '::' | '0.0.0.0' | string
+  webroot?: string
   proxyHost?: string
   auth: 'form' | 'none'
   theme: string